lmtable  filter2: converted some int into long.
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk@901 1f5c12ca-751b-0410-a591-d2e778427230

+++ irstlm/src/lmtable.h
@@ -69,7 +69,7 @@ private:
 public:
     
 	disktable(std::fstream& inp, int buffersize, int entrysize,long max_entries){
-       buffer_size=buffersize;
+      buffer_size=buffersize;
       entry_size=entrysize;
       buffer=new char[(buffer_size+1) * entry_size];
       current_border=0;
@@ -79,7 +79,7 @@ public:
   
   ~disktable() {delete [] buffer;};
   
-  char* get(std::fstream& inp,int position){    
+  char* get(std::fstream& inp,long position){    
     assert(position < file_border);
     
     //you can look back at maximum one position before the first in the buffer!
